requirements_build.txt
# MediVote Professional Installer Build Requirements
# Dependencies needed to build the professional installer

# Core application dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# GUI dependencies for desktop application
tkinter-tooltip>=2.2.0
pillow>=10.0.0

# Build and packaging dependencies
pyinstaller>=5.13.0
auto-py-to-exe>=2.40.0
cx-freeze>=6.15.0
nuitka>=1.8.0

# Icon and image processing
pillow>=10.0.0

# Windows-specific dependencies
pywin32>=306; sys_platform == "win32"
pywin32-ctypes>=0.2.2; sys_platform == "win32"

# macOS-specific dependencies
py2app>=0.28.0; sys_platform == "darwin"
dmgbuild>=1.6.0; sys_platform == "darwin"

# Linux-specific dependencies

# Additional utilities
wheel>=0.41.0
setuptools>=68.0.0
build>=0.10.0
twine>=4.0.0 